Rename operations/formats.py -> channels.py
"formats" was too generic to remember. Everything in the module is about
channels: addressing (col_to_board_chan), identifiers (get_row_col), and
dead-channel masks (make/write/read_dead_masks). channels.py names the
unifying theme.

git mv (history preserved); updated the 4 importers (session,
unit_conversions, analysis, __init__), module header, __all__ comment,
SOFTWARE_GUIDE, and current-state G9/dead-mask PLAN references. Public
export names unchanged -> no API break. Verified in warm-tdm-r615: 42
exports intact, channels imports, old name gone.
